+++ build.gradle
@@ -9,7 +9,6 @@ repositories {
 }
 
 dependencies {
-    implementation 'junit:junit:4.13.1'
     testImplementation 'org.assertj:assertj-core:3.22.0'
     testImplementation 'org.junit.jupiter:junit-jupiter:5.8.2'
 }
