+++ src/main/java/racingcar/Application.java
@@ -3,7 +3,7 @@ package racingcar;
 import racingcar.controller.MainController;
 
 public class Application {
-    public static void main(String[] args) {
+    public static void main(final String... args) {
         MainController mainController = new MainController();
         mainController.play();
     }
