+++ src/test/java/racingcar/controller/ValidationTest.java
@@ -3,6 +3,7 @@ package racingcar.controller;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import racingcar.util.Validation;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
